+++ src/Components/Contact/Contact.jsx
@@ -26,8 +26,8 @@ const Contact = () => {
                     }
                 ></SharedBanner> 
 
-                <div className="bg-cover bg-center bg-no-repeat ps-7" style={{ backgroundImage: `url(${bg})` }}>
-                    <div className="md:flex justify-center mx-auto pb-8 pt-8">
+                <div className="bg-cover bg-center bg-no-repeat" style={{ backgroundImage: `url(${bg})` }}>
+                    <div className="md:flex justify-center gap-4 mx-auto pb-8 pt-8">
                         <div className="text-neutral">
                             <h1 className="text-3xl text-[#F99F24]">Address</h1>
                             <div className="flex items-center gap-3 mt-5">
